feat(vm): support selecting the device for the CloudInit image
Added an `interface` entry to the `initialization` block which
controls which IDE device the CloudInit image is attached to. If the
setting is not present, autodetection will occur, so cloning a template
with a non-default CloudInit drive will work. If no `interface` setting
nor existing image is present, the default `ide2` will be used.