fix(vm): Fixed missing default for disk discard
The default value of "ignore" was missing from the disk discard
parameter of the terraform module. This commit sets the default to
"ignore" instead of the current "".

This fixes terraform incorrectly flagging that the state is different than
what is currently running on the proxmox node when the discard option isn't
defined under the disk object.

Signed-off-by: Daniel Lowry <development@daniellowry.co.uk>