chore(docs): add 'stop_on_destroy' configuration to cloud image examples
Added the 'stop_on_destroy' parameter to the Proxmox VM configurations in the cloud image documentation and example files for CentOS, Ubuntu, and Debian. This parameter should be set to true if the QEMU agent is not installed or enabled on the VM, enhancing clarity for users configuring their virtual machines.

Signed-off-by: Pavel Boldyrev <627562+bpg@users.noreply.github.com>