fix(lxc): delete dns attributes when set to null
As it seems, to force a container to use host settings, it is not enough to set previous values ('nameserver' and/or 'searchdomain') to empty strings. These must be deleted using 'delete' attribute of the 'PUT /nodes/{node}/lxc/{vmid}/config' request.

One particular case is also handled. When a lxc container has 'use host settings' for 'nameserver' and/or 'searchdomain', the Proxmox API does not return the actual values used. Not only does it not allow to make these fields 'Computed', but also creates the following problem: when there was no 'dns' block, and then one is added to with 'domain = null' and 'servers = null', the plan reports that an empty 'dns' block is added. And this diff is perpetual, meaning that any consecutive 'terraform plan/apply' will report this change. To handle this case, a custom 'DiffSuppressFunc'.

Also, this commit removes 'MinItems: 0' from dns resource as it is 'Optional' (in this case 'MinItems' is ignored).

Signed-off-by: Stanislav Shamilov <shamilovstas@protonmail.com>