fix(vm): correctly detect pool membership for VM
This makes sure that whenever a VM's state is read,
the provider also checks the API to determine in
which pool the VM actually is, not just according to
the saved state. The API unfortunately doesn't provide
the VM's pool in the VM's state, so the only solution
is to check every pool if a member is the VM we are
looking for.

Signed-off-by: TobiPeterG <github.threefold020@passmail.net>