fix(lxc): race condition in container clone / reboot operations (#2320)
* disable parallelism for lxc tests
add waiting for status "running" after update that triggered reboot

Signed-off-by: Stanislav Shamilov <shamilovstas@protonmail.com>

* fix(pool): refactoring of lxc acceptance tests

It seems that some tests running in parallel and using the same container were causing some of the lock timeouts on the proxmox host used for testing. This commit splits a single test, parts of which were run in parallel and used the same container, into several independent tests using their own containers. This should reduce the overall flakiness of the tests.
Besides, a few tests were fixed. The tests for ipv4/ipv6 were using linux bridge "vmbr1", which wasn't created beforehand. Also, removed some of the WaitForContainerStatus calls done after updating the container since previous commit introduced a wait in the `updateContainer` method which waits for the container to be rebooted.
Also, the template used for running containers was changed from Ubuntu 24.04 to Alpine 3.22. This makes the downloading of the template faster as the Alpine template is much smaller.

Signed-off-by: Stanislav Shamilov <shamilovstas@protonmail.com>

* fix(lxc): race condition in container clone / reboot operations

Signed-off-by: Pavel Boldyrev <627562+bpg@users.noreply.github.com>

* linter

Signed-off-by: Pavel Boldyrev <627562+bpg@users.noreply.github.com>

---------

Signed-off-by: Stanislav Shamilov <shamilovstas@protonmail.com>
Signed-off-by: Pavel Boldyrev <627562+bpg@users.noreply.github.com>
Co-authored-by: Pavel Boldyrev <627562+bpg@users.noreply.github.com>